Load template routes into the TemplateRoutingService.
